chore: rename `hot.ts` to `hmr.ts`
