+++ packages/types/src/index.ts
@@ -3,4 +3,4 @@ export * from './setups'
 export * from './config'
 export * from './toc'
 export * from './cli'
-export * from './hot'
+export * from './hmr'
